Extract `CachingEFRepository` out of `EFRepository`
