+++ src/Domain/LeanCode.DomainModels.EF/EFRepository.cs
@@ -1,10 +1,7 @@
-using System.Diagnostics.CodeAnalysis;
-using System.Numerics;
 using LeanCode.DomainModels.DataAccess;
 using LeanCode.DomainModels.Model;
 using LeanCode.TimeProvider;
 using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Internal;
 
 namespace LeanCode.DomainModels.EF;
 
@@ -72,41 +69,5 @@ public abstract class EFRepository<TEntity, TIdentity, TContext> : IRepository<T
     public virtual async Task<TEntity?> FindAsync(TIdentity id, CancellationToken cancellationToken = default)
     {
         return await BaseQuery().AsTracking().FirstOrDefaultAsync(e => e.Id.Equals(id), cancellationToken);
-    }
-
-    [SuppressMessage(
-        "?",
-        "EF1001",
-        Justification = "This is basically the `EntityFinder.FindTracked` to mimic `FindAsync` snapshot re-use."
-    )]
-    protected TEntity? FindTracked(TIdentity id)
-    {
-        // Safety: aggregates are bound to have Id as a primary key by design.
-        var primaryKey = DbContext.Model.FindEntityType(typeof(TEntity))!.FindPrimaryKey()!;
-        return ((IDbContextDependencies)DbContext).StateManager!.TryGetEntryTyped(primaryKey, id)?.Entity as TEntity;
-    }
-
-    [SuppressMessage(
-        "?",
-        "EF1001",
-        Justification = "This is basically the `EntityFinder.FindTracked` to mimic `FindAsync` snapshot re-use."
-    )]
-    protected TEntity? FindTracked(params object[] id)
-    {
-        // Safety: aggregates are bound to have Id as a primary key by design.
-        var primaryKey = DbContext.Model.FindEntityType(typeof(TEntity))!.FindPrimaryKey()!;
-        return ((IDbContextDependencies)DbContext).StateManager!.TryGetEntryTyped(primaryKey, id)?.Entity as TEntity;
-    }
-
-    protected ValueTask<TEntity?> FindTrackedOrLoadNewAsync(TIdentity id, Func<DbSet<TEntity>, Task<TEntity?>> query)
-    {
-        var tracked = FindTracked(id);
-        return tracked is not null ? new(tracked) : new(query(DbSet));
-    }
-
-    protected ValueTask<TEntity?> FindTrackedOrLoadNewAsync(object[] id, Func<DbSet<TEntity>, Task<TEntity?>> query)
-    {
-        var tracked = FindTracked(id);
-        return tracked is not null ? new(tracked) : new(query(DbSet));
     }
 }
